# Fixed task distribution in the USA model (100x speedup)