--- a/openquake/hazardlib/sourceconverter.py
+++ b/openquake/hazardlib/sourceconverter.py
@@ -298,8 +298,11 @@ class SourceGroup(collections.abc.Sequence):
             else:
                 sources.append(src)
         out = []
-        for block in block_splitter(
-                sources, maxweight, operator.attrgetter('num_ruptures')):
+        def weight(src):
+            if src.code == b'F':  # consider it much heavier
+                return src.num_ruptures * 5
+            return src.num_ruptures
+        for block in block_splitter(sources, maxweight, weight):
             sg = copy.copy(self)
             sg.sources = block
             out.append(sg)