With respect to yesterday the time to generate the ruptures went down from 10,000 seconds to 100 seconds and the slowest task went down from 9_004s to 77s (117x speedup).